task6指摘対応

+++ twitter/views.py
@@ -11,6 +11,7 @@ from django.views import generic
 
 from .models import Tweet,Follow,FavoriteTweet
 from .forms import TweetForm
+from itertools import chain
 
 # Create your views here.
 #ツイートの一覧表示
@@ -21,6 +22,7 @@ def top(request):
     return render(request, 'twitter/top.html', {'tweet_list': tweet_list})
 
 #他の人のツイート（自分含む）
+@login_required
 def tweet_user(request, tweet_user_id):
     tweet_user = get_object_or_404(User, id=tweet_user_id)
     tweet_list = Tweet.objects.filter(user=tweet_user).order_by('created_time').reverse()[:20]
@@ -28,6 +30,7 @@ def tweet_user(request, tweet_user_id):
     return render(request, 'twitter/tweet_list.html', {'tweet_user': tweet_user, 'tweet_list': tweet_list, 'is_follow': is_follow})
 
 #フォロー機能
+@login_required
 def user_follow(request, tweet_user_id):
     tweet_user = get_object_or_404(User, id=tweet_user_id)
     follow, created = Follow.objects.get_or_create(
@@ -39,6 +42,7 @@ def user_follow(request, tweet_user_id):
     return redirect(request.META.get('HTTP_REFERER'))
 
 #フォロー削除機能
+@login_required
 def follow_deleted(request, tweet_user_id):
     tweet_user = get_object_or_404(User, id=tweet_user_id)
     follow = Follow.objects.filter(follower=request.user, followee=tweet_user)
@@ -46,12 +50,14 @@ def follow_deleted(request, tweet_user_id):
     return redirect(request.META.get('HTTP_REFERER'))
 
 #お気に入り一覧
+@login_required
 def favorite_list(request):
     favorite_tweet_list = FavoriteTweet.objects.filter(user=request.user).values_list('tweet', flat=True)
     tweet_list = Tweet.objects.filter(id__in=favorite_tweet_list).order_by('created_time').reverse()[:20]
     return render(request, 'twitter/favorite.html', {'tweet_list': tweet_list})
 
 #お気に入り追加機能
+@login_required
 def add_favorite(request, favorite_tweet_id):
     favorite_tweet = get_object_or_404(Tweet, id=favorite_tweet_id)
     tweet, created = FavoriteTweet.objects.get_or_create(
@@ -63,10 +69,12 @@ def add_favorite(request, favorite_tweet_id):
     return redirect(request.META.get('HTTP_REFERER'))
 
 #お気に入り削除機能
+@login_required
 def del_favorite(request, favorite_tweet_id):
     favorite_tweet = get_object_or_404(Tweet, id=favorite_tweet_id)
     tweet = FavoriteTweet.objects.filter(user=request.user, tweet=favorite_tweet)
-    tweet.delete()
+    if tweet.exists():
+        tweet.delete()
     return redirect(request.META.get('HTTP_REFERER'))
 
 #ツイート処理
